refactor: Ordenar las clase de excepcion personalizadas en carpetas para mantener un orden
