+++ src/main/java/com/euphony/streaming/service/implementation/PlaylistServiceImpl.java
@@ -4,7 +4,11 @@ import com.euphony.streaming.dto.request.PlaylistRequestDTO;
 import com.euphony.streaming.dto.response.PlaylistResponseDTO;
 import com.euphony.streaming.entity.PlaylistEntity;
 import com.euphony.streaming.entity.UsuarioEntity;
-import com.euphony.streaming.exception.custom.*;
+import com.euphony.streaming.exception.custom.playlist.PlaylistCreationException;
+import com.euphony.streaming.exception.custom.playlist.PlaylistDeletionException;
+import com.euphony.streaming.exception.custom.playlist.PlaylistNotFoundException;
+import com.euphony.streaming.exception.custom.playlist.PlaylistUpdateException;
+import com.euphony.streaming.exception.custom.user.UserNotFoundException;
 import com.euphony.streaming.repository.PlaylistRepository;
 import com.euphony.streaming.repository.UsuarioRepository;
 import com.euphony.streaming.service.interfaces.IPlaylistService;
