+++ src/main/java/com/euphony/streaming/service/implementation/ProfileUserServiceImpl.java
@@ -3,8 +3,8 @@ package com.euphony.streaming.service.implementation;
 import com.euphony.streaming.dto.request.UserProfileRequestDTO;
 import com.euphony.streaming.dto.response.UserProfileResponseDTO;
 import com.euphony.streaming.entity.PerfilUsuarioEntity;
-import com.euphony.streaming.exception.custom.UserNotFoundException;
-import com.euphony.streaming.exception.custom.UserUpdateException;
+import com.euphony.streaming.exception.custom.user.UserNotFoundException;
+import com.euphony.streaming.exception.custom.user.UserUpdateException;
 import com.euphony.streaming.publisher.UserProfilePublisher;
 import com.euphony.streaming.repository.PerfilUsuarioRepository;
 import com.euphony.streaming.service.interfaces.IProfileUserService;
