+++ src/main/java/com/euphony/streaming/service/implementation/UserServiceImpl.java
@@ -4,8 +4,8 @@ import com.euphony.streaming.dto.request.UserRequestDTO;
 import com.euphony.streaming.dto.response.UserResponseDTO;
 import com.euphony.streaming.entity.RolEntity;
 import com.euphony.streaming.entity.UsuarioEntity;
-import com.euphony.streaming.exception.custom.UserCreationException;
-import com.euphony.streaming.exception.custom.UserNotFoundException;
+import com.euphony.streaming.exception.custom.user.UserCreationException;
+import com.euphony.streaming.exception.custom.user.UserNotFoundException;
 import com.euphony.streaming.repository.RolRepository;
 import com.euphony.streaming.repository.UsuarioRepository;
 import com.euphony.streaming.service.interfaces.IUserService;
